Create Cental_device_code
This is the code I have created for the Central device (Nano 33 BLE SENSE). This is still a work in progress. This code does compile but it still needs to be tested to see how this code works with a peripheral. I will also need to see how if I have used the correct BLECharacteristic function for this project. I will also need to adjust how the gesture button is going to work (my goal is to press and hold the button while performing the gesture so that the trajectory can be recognized)