Updated Central_device_code
I had added the necessary code to send the command to light up the LEDs from the peripheral. I am currently waiting the arrival of the other Nano 33 BLE to test both codes. This code still needs to be tested. My next step that will be to to incorporate the motion sensing programming to this code and how the buttons should work.