Updated Peripheral_device_code
This update version of the code compiles. This code lights up LEDs on the peripheral based on the inputs of the Central device. I am currently waiting to obtain my other Nano 33 in order to test both Peripheral_device_code and Central_device_code (https://github.com/MSU-Drone-Project/Motion_sensing_for_drone_control/blob/master/Cental_device_code). This is still an on going project. This is not complete as my main goal is to have gestures recognized from the central.